requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.5
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-import-export==4.3.9
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-schema-viewer==0.5.3
django-tinymce==4.1.0
djangorestframework==3.16.1
docutils==0.22
joblib==1.5.1
numpy==2.3.2
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
threadpoolctl==3.6.0
tzdata==2025.2
